Switch to flannel v0.13.1 RC (fork)
Switch to flannel v0.13.1 RC. This is still a fork but the plan is to
close the gap with upstream flannel by v0.13.1 release and no longer use
a fork.

Includes a few strict compatibility changes on k3s side due to flannel
API changes:

* ba8c21659ec52c4d522eebb6ab92e8c6cf5aec4c
  upgrade client-go to 1.19.4

* 117c10277ed6561f7c221865e16b3c2b17459866
  pass waitgroup pointer

* d31b0dc85a5a15bda5e606acbbbb9f7089441a87
  windows: replace old netsh (rakelkar/gonetsh) with powershell commands

Includes all vendoring changes as well.